chore: Move helper function out from the views file to a new helpers file
